RESOLVED - # 1308: Cannot Parse Inf as a value from MATLAB State Files https://mantis.sjrx.net/view.php?id=1308 RESOLVED - # 1307 : Allow out of order runs in the runHistory File

+++ src/ca/ubc/cs/beta/smac/ac/runs/ExistingAlgorithmRun.java
@@ -36,7 +36,7 @@ public class ExistingAlgorithmRun extends AbstractAlgorithmRun {
 			this.acResult = RunResult.getAutomaticConfiguratorResultForCode(Integer.valueOf(resultLine[0]));
 			this.runtime = Double.valueOf(resultLine[1]);
 			this.runLength = Integer.valueOf(resultLine[2]);
-			this.quality = Integer.valueOf(resultLine[3]);
+			this.quality = Double.valueOf(resultLine[3]);
 			this.resultSeed = Long.valueOf(resultLine[4]);
 			this.resultLine = result;
